refactor execute entrypoints into specific files
