+++ contracts/governance_admin/src/lib.rs
@@ -3,5 +3,6 @@ extern crate cosmwasm_std;
 
 pub mod contract;
 pub mod error;
+pub mod execute;
 pub mod state;
 pub mod utils;
