+++ contracts/governance_admin/tests/dexter-create-pool.rs
@@ -1,27 +1,23 @@
-use cosmwasm_std::{Addr, Coin, to_binary, Uint128, Decimal, QueryRequest};
-use dexter::{governance_admin, vault::{FeeInfo, NativeAssetPrecisionInfo}, asset::Asset};
+use cosmwasm_std::{Addr, Coin, Uint128};
+use dexter::{asset::Asset, vault::FeeInfo};
 use persistence_std::types::cosmos::gov::v1::QueryParamsRequest;
-use weighted_pool::state::WeightedParams;
 
 mod utils;
 
 #[test]
 fn test_basic_functions() {
-
     let governance_params_query = QueryParamsRequest {
         params_type: String::from("deposit"),
     };
 
     println!("{:?}", governance_params_query);
-
 }
 
 #[test]
 fn test_create_pool() {
-
     let vault_creator: Addr = Addr::unchecked("vault_creator".to_string());
     let keeper_owner: Addr = Addr::unchecked("keeper_owner".to_string());
-    let alice_address: Addr = Addr::unchecked("alice".to_string());
+    let _alice_address: Addr = Addr::unchecked("alice".to_string());
 
     let mut app = utils::mock_app(
         vault_creator.clone(),
@@ -42,13 +38,10 @@ fn test_create_pool() {
         protocol_fee_percent: 20,
     };
 
-    let (vault_addr, keeper_addr, governance_admin) = utils::instantiate_contracts(&mut app, 
-        &vault_creator, 
-        &keeper_owner,
-        fee_info
-    );
+    let (_vault_addr, _keeper_addr, _governance_admin) =
+        utils::instantiate_contracts(&mut app, &vault_creator, &keeper_owner, fee_info);
 
-    let asset_infos_with_weights = vec![
+    let _asset_infos_with_weights = vec![
         Asset::new_native("uxprt".to_string(), Uint128::from(1u128)),
         Asset::new_native("uatom".to_string(), Uint128::from(1u128)),
     ];
@@ -57,11 +50,11 @@ fn test_create_pool() {
     // Currently, we'd use an account that doesn't have a private key in tests directly.
     // Ideally, during the actual run, the users trigger the Execute contract governance flow which triggers it from that address
     // but on pool creator's behalf
-    // let msg = dexter::governance_admin::ExecuteMsg::CreateNewPool { 
-    //     vault_addr: vault_addr.to_string(), 
-    //     bootstrapping_amount_payer: vault_creator.to_string(), 
-    //     pool_type: dexter::vault::PoolType::Weighted {}, 
-    //     fee_info: None, 
+    // let msg = dexter::governance_admin::ExecuteMsg::CreateNewPool {
+    //     vault_addr: vault_addr.to_string(),
+    //     bootstrapping_amount_payer: vault_creator.to_string(),
+    //     pool_type: dexter::vault::PoolType::Weighted {},
+    //     fee_info: None,
     //     native_asset_precisions: vec![
     //         NativeAssetPrecisionInfo {
     //             denom: "uxprt".to_string(),
@@ -87,6 +80,4 @@ fn test_create_pool() {
 
     // send the message along with funds. for now, they'd be directly from the non private key account
     // but ideally, funds will be sent with the governance execute contract message
-    
-
 }
